Get rid of implementation specific workspace information
